+++ CHANGELOG.md
@@ -377,6 +377,7 @@
 * IControlPlayerCamera now provides getter for player audio listener
 * Fixed 2D collision matrix not matching 3D collision matrix
 * Added PathFollower entity that pushes actors along a path of NavigationNodes
+* Added ThingPusher entity that pushes other things along a path of NavigationNodes
 * Added conditional/microscript resolvers for PlayerFlags and SessionFlags (conditional only) [Experimental]
 * Added ScriptStringSubber that calls scripts and substitutes the result [Experimental]
 * DelayScaled can now be used from threads other than the main thread [Untested]